feat: 1.Update: rename ThrowableRunnable.
